refactor: :recycle: HitterQuizState を別ファイルに切り出し
