+++ package/model/lib/src/feature/quiz/quiz.dart
@@ -3,6 +3,7 @@ export 'application/hitter_quiz_service.dart';
 export 'application/hitter_quiz_state_provider.dart';
 export 'application/share_quiz_result_service.dart';
 export 'domain/hitter.dart';
+export 'domain/hitter_quiz.dart';
 export 'domain/hitter_quiz_state.dart';
 export 'domain/stats_value.dart';
 export 'infrastructure/entity/hitting_stats.dart';
